[댓글 기능] 이지 미션 제출합니다 (#143)
* [DOCS] README.md 업데이트

* [FEAT] User의 PK를 Id로 교체

* [FIX] 드랍다운 버그 수정

* [FEAT] 작성자만 글 수정가능하게 제한

* [FEAT] 댓글 생성 조회 기능 추가

* [FEAT] 댓글 수정, 삭제 버튼 추가

* [FEAT]댓글 업데이트 기능 추가

* [FEAT] 부모 엔티티 삭제시 자식 엔티티 삭제 기능 추가

* [FIX] 로그아웃 버그 수정

* [FEAT] 댓글 테스트 추가

* [REFACTOR] mypage edit 수정, 필요없는 파일 삭제

* [REFACTOR] ArticleController 기존 테스트 에러 수정

* [REFACTOR] Mockito 사용해서 UserService Test 수정

* [TEST] ArticleService 테스트 코드 작성

* [TEST] CommentService 테스트 코드 작성

- ArticleServiceTests 에서 UserRepository를 Mock으로 생성하던 것을 UserService를 Mock으로 생성하는 것으로 변경
- 서비스가 Session에서 로그인 정보를 빼오는 것을 컨트롤러에서 수행하도록 변경

* [TEST] CommentController 테스트 코드 작성

* [REFACTOR] DTO 변경

- DTO에서 Entity를 필드로 가지지 않도록 수정
- Converter 클래스를 인터페이스로 추출
- UserDto를 ResponseDto와 RequestDto로 구분
- Converter에서 DTO의 타입을 다르게 하기위해 제너릭을 3개 사용

* [REFACTOR] Test 코드 에러 수정

* [FEAT] 비로그인시, 혹은 로그인시 접속 가능한 페이지 처리

* [FEAT] 댓글의 작성자만 수정, 삭제를 할 수 있는 기능 추가

* [REFACTOR] 테스트 에러 수정

* [REFACTOR] Session에서 Email 대신 User를 가지도록 수정

* [REFACTOR] 공통되는 부분 ControllerTests의 추상 메서드로 추출

* [REFACTOR] Converter를 삭제하고 Assembler로 변경

- DTO와 Entity를 서로 바꿔주는 Converter 삭제
- Entity에서 DTO를 만들어주는 Assembler 생성
- DTO와 Entity 생성자 변경, Entity Builder 패턴 적용

* [REFACTOR] Comment 객체가 작성자인지 판단하도록 리팩토링

- Comment에게 User를 보내 작성자인지 확인하도록 수정
- Comment 수정 시에 Domain에서 작성자 판단 및 내용을 수정하도록 변경

* [REFACTOR] Entity도 Mock을 사용하도록 리팩토링

* [FEAT] Article 수정, 삭제에서도 작성자만 가능하도록 수정

- 세션의 값과 비교해서 작성자인지 확인하도록 하는 기능 추가

* [TEST] Controller, Domain 테스트 추가

- Interceptor에서 작업이 모두 처리되기 전에 먼저 redirect 되는 에러 수정

+++ src/main/java/techcourse/myblog/application/service/ArticleService.java
@@ -1,54 +1,93 @@
 package techcourse.myblog.application.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.application.converter.ArticleConverter;
+import techcourse.myblog.application.assembler.ArticleAssembler;
 import techcourse.myblog.application.dto.ArticleDto;
 import techcourse.myblog.application.service.exception.NotExistArticleIdException;
+import techcourse.myblog.application.service.exception.NotMatchArticleAuthorException;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.ArticleRepository;
+import techcourse.myblog.domain.User;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ArticleService {
-    private final ArticleConverter articleConverter = ArticleConverter.getInstance();
+    private static final Logger log = LoggerFactory.getLogger(ArticleService.class);
+
     private final ArticleRepository articleRepository;
+    private final UserService userService;
+
+    private final ArticleAssembler articleAssembler = ArticleAssembler.getInstance();
 
     @Autowired
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, UserService userService) {
         this.articleRepository = articleRepository;
+        this.userService = userService;
     }
 
     @Transactional
-    public Long save(ArticleDto articleDto) {
-        return articleRepository.save(articleConverter.convertFromDto(articleDto)).getId();
+    public Long save(ArticleDto articleDto, User user) {
+        User author = userService.findUserById(user.getId());
+        Article article = new Article.ArticleBuilder()
+                .contents(articleDto.getContents())
+                .title(articleDto.getTitle())
+                .coverUrl(articleDto.getCoverUrl())
+                .author(author)
+                .build();
+
+        log.info(article.toString());
+        return articleRepository.save(article).getId();
+    }
+
+    Article findArticleById(Long articleId) {
+        return articleRepository.findById(articleId)
+                .orElseThrow(() -> new NotExistArticleIdException("존재하지 않는 Article 입니다."));
     }
 
     @Transactional(readOnly = true)
     public ArticleDto findById(Long articleId) {
-        Article article = articleRepository.findById(articleId)
-                .orElseThrow(() -> new NotExistArticleIdException("존재하지 않는 Article 입니다."));
-
-        return articleConverter.convertFromEntity(article);
+        Article article = findArticleById(articleId);
+        return articleAssembler.convertEntityToDto(article);
     }
 
     @Transactional
-    public void removeById(Long articleId) {
+    public void removeById(Long articleId, User user) {
+        Article article = articleRepository.findById(articleId)
+                .orElseThrow(() -> new NotExistArticleIdException("해당 Id의 글이 존재하지 않습니다."));
+        if (article.isNotAuthor(user)) {
+            throw new NotMatchArticleAuthorException("너는 이 글에 작성자가 아니다. 꺼져라!");
+        }
         articleRepository.deleteById(articleId);
     }
 
     @Transactional
-    public void modify(Long articleId, ArticleDto articleDto) {
+    public void modify(Long articleId, ArticleDto articleDto, User user) {
         Article article = articleRepository.findById(articleId)
-                .orElseThrow(() -> new NotExistArticleIdException(""));
-
-        article.modify(articleConverter.convertFromDto(articleDto));
+                .orElseThrow(() -> new NotExistArticleIdException("해당 Id의 글이 존재하지 않습니다."));
+        article.modify(articleDto, user);
     }
 
     @Transactional(readOnly = true)
     public List<ArticleDto> findAll() {
-        return articleConverter.createFromEntities(articleRepository.findAll());
+        return articleRepository.findAll().stream()
+                .map(articleAssembler::convertEntityToDto)
+                .collect(Collectors.toList());
+    }
+
+    public void matchAuthor(Long articleId, User user) {
+        User author = findArticleById(articleId).getAuthor();
+        if (!author.equals(user)) {
+            throw new NotMatchArticleAuthorException("너는 이 글에 작성자가 아니다. 꺼져라!");
+        }
+    }
+
+    public boolean matchAuthor(ArticleDto articleDto, User user) {
+        return articleDto.getAuthor().match(user);
     }
 }
